+++ semcore/feature-highlight/src/components/notice/Notice.type.ts
@@ -1,0 +1,4 @@
+import type { Intergalactic } from '@semcore/core';
+import type { NoticeSmartProps } from '@semcore/notice';
+
+export type HighlightedNoticeComponent = Intergalactic.Component<'div', NoticeSmartProps, {}>;