+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.1] - 2025-10-17
+
+### Fixed
+
+- Limited mode doesn't prevent interaction with accordion.
+
 ## [16.4.0] - 2025-10-03
 
 ### Added